requirements.txt
Flask==0.10.1

# Application server for both development and production.
gunicorn==19.4.5
werkzeug==0.14.1

# Environment variables.
python-decouple==3.3

# Testing and static analysis.
pytest==5.1.0
pytest-cov==2.10.1
flake8==3.7.8

# CLI.
Click==6.4

# Data and workers.
psycopg2-binary==2.8.6
Flask-SQLAlchemy==2.1
SQLAlchemy==1.2.0
redis==3.3.7
celery==4.4.0

# Forms.
Flask-WTF==0.14.2
WTForms-Components==0.10.4
WTForms-Alchemy==0.16.9

# Extensions.
flask-debugtoolbar==0.10.0
Flask-Mail==0.9.1
Flask-Login==0.3.2